# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Furniture ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: roleId
          in: query
          schema:
            type: integer
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"

    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /users/{id}/toggle-active:
    patch:
      operationId: toggleUserActive
      tags: [users]
      summary: Activate or deactivate user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset user password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset

  # ROLES
...
components:
  schemas:
...
    ResetPasswordBody:
      type: object
      properties:
        newPassword:
          type: string
      required: [newPassword]

    # USERS
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        email:
          type: string
          nullable: true
        roleId:
          type: integer
        role:
          $ref: "#/components/schemas/Role"
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, mobile, roleId, isActive, createdAt]
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        email:
          type: string
          nullable: true
        password:
          type: string
        roleId:
          type: integer
      required: [name, mobile, password, roleId]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        email:
          type: string
          nullable: true
        roleId:
          type: integer
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # ROLES
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        permissions:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PermissionSet"
        createdAt:
          type: string
          format: date-time
      required: [id, name, permissions, createdAt]
    PermissionSet:
      type: object
      properties:
        view:
          type: boolean
        add:
          type: boolean
        edit:
          type: boolean
        delete:
          type: boolean
      required: [view, add, edit, delete]
...